Moved form.js route into its own folder, added its own md file
